# mydiplom/openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
  /api/v1/partner/state/:
    get:
      operationId: PartnerState
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                  next:
                    type: string
                    nullable: true
                  previous:
                    type: string
                    nullable: true
                  results:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          readOnly: true
                        name:
                          type: string
                          maxLength: 100
                        state:
                          type: boolean
                      required:
                      - name
          description: ''
      headers:
        Authorization:
          type: string
    post:
      operationId: PartnerState
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                state:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    readOnly: true
                  name:
                    type: string
                    maxLength: 100
                  state:
                    type: boolean
                required:
                - name
          description: ''
      headers:
        Authorization:
          type: string
  /api/v1/partner/state/{id}/:
    get:
      operationId: PartnerState
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    readOnly: true
                  name:
                    type: string
                    maxLength: 100
                  state:
                    type: boolean
                required:
                - name
          description: ''
      headers:
        Authorization:
          type: string
    put:
      operationId: PartnerState
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                state:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    readOnly: true
                  name:
                    type: string
                    maxLength: 100
                  state:
                    type: boolean
                required:
                - name
          description: ''
      headers:
        Authorization:
          type: string
    patch:
      operationId: PartnerState
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                state:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    readOnly: true
                  name:
                    type: string
                    maxLength: 100
                  state:
                    type: boolean
                required:
                - name
          description: ''
      headers:
        Authorization:
          type: string
    delete:
      operationId: PartnerState
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '204':
          description: ''
      headers:
        Authorization:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                state:
                  type: string
  /api/v1/partner/orders:
    get:
      operationId: PartnerOrders
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: {}
          description: ''
      headers:
        Authorization:
          type: string
  /api/v1/user/details:
    get:
      operationId: AccountDetails
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: {}
          description: ''
      headers:
        Authorization:
          type: string
    post:
      operationId: AccountDetails
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: ''
      headers:
        Authorization:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                company:
                  type: string
                position:
                  type: string
                type:
                  type: string
...
  /api/v1/categories:
    get:
      operationId: CategoryView
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                  next:
                    type: string
                    nullable: true
                  previous:
                    type: string
                    nullable: true
                  results:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          readOnly: true
                        name:
                          type: string
                          maxLength: 50
                      required:
                      - name
          description: ''
      headers:
        Authorization:
          type: string
  /api/v1/shops:
    get:
      operationId: ShopView
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                  next:
                    type: string
                    nullable: true
                  previous:
                    type: string
                    nullable: true
                  results:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          readOnly: true
                        name:
                          type: string
                          maxLength: 100
                        state:
                          type: boolean
                      required:
                      - name
          description: ''
      headers:
        Authorization:
          type: string
  /api/v1/products:
    get:
      operationId: ProductInfoView
      description: ''
      parameters:
      - name: shop_id
        in: query
        required: false
        description: Output format
        schema:
          type: string
      - name: category_id
        in: query
        required: false
        description: Output format
        schema:
          type: string
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: {}
          description: ''
      headers:
        Authorization:
          type: string
...
  /api/v1/partner/update:
    post:
      operationId: PartnerUpdate
      description: ''
      parameters:
      - name: Authorization
        in: header
        required: false
        description: Output format
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: ''
      headers:
        Authorization:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                url:
                  type: string
...